Add files via upload
The problem with the current parking system is that the driver doesn’t know
whether a slot is empty or not in a parking space and also a lot of times drivers
end up bumping the cars due to disturbances or chaos in the parking area. With
the increase in the global human population, which implies increase in number of
vehicles has only aggravated the whole issues of parking slots and vehicle
safety.
so we have implemented our own parking assistance system which tells the user whether there is an empty slot and an automated entry and exit barrier.